Refactor: improve settings folder
